make changes to directory layout
