--- src/preprocess/.gitignore
@@ -1,3 +1,0 @@
-/stocks
-Preprocessing.ipynb
-/.ipynb_checkpoints
