--- src/preprocess/merge.py
@@ -1,21 +1,0 @@
-from common import create_data
-import os
-
-# main_dir = os.getcwd()
-
-# Concatenating stock data
-
-# STOCKS
-# os.chdir(main_dir)
-# os.chdir("stocks/Data/Stocks")
-# stock_list = os.listdir()
-# stocks = create_data(stock_list)
-
-# Concatenating ETF data
-
-# ETFs
-# os.chdir(main_dir)
-# os.chdir("stocks/Data/ETFs")
-# etf_list = os.listdir()
-
-# etf = create_data(etf_list)
